Removes flask_login, adds flask_jwt user loading

+++ app/products/routes.py
@@ -4,37 +4,31 @@ from . import (
     products as products_bp,
     db,
     Product,
+    seller_required
 )
 from flask import (
     jsonify,
     abort,
     request
 )
-from flask_login import (
-    current_user,
-    login_required
-)
 from flask_jwt_extended import (
-    jwt_required
+    jwt_required,
+    current_user
 )
 from sqlalchemy import select
 
 @products_bp.get('/product/<int:product_id>')
-@login_required
-@jwt_required()
 def getProduct(product_id):
-    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar()
+    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar_one_or_none()
 
     try:
-        return jsonify({
-            "product": product
-        })
+        return jsonify(product=product)
     except:
-        return "Error occurred", 500
+        return jsonify(msg="Error occurred!"), 500
 
 @products_bp.post('/create')
-@login_required
 @jwt_required()
+@seller_required()
 def createProduct():
     try:
         data = request.get_json()
@@ -44,39 +38,43 @@ def createProduct():
             price=data['price'],
             quantity=data['quantity'],
             description=data['description'],
-            user_id=data['user_id']
+            user_id=current_user.id
         )
 
         db.session.add(product)
         db.session.commit()
 
-        return jsonify({'result': True}), 201
+        return jsonify(msg="Succesfully created product!"), 201
     except:
         db.session.rollback()
-        return "Error occurred", 500
+        return jsonify(msg="Error occurred!"), 500
     
 @products_bp.delete('/delete/<int:product_id>')
-@login_required
 @jwt_required()
 def deleteProduct(product_id):
-    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar()
+    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar_one_or_none()
     
     try:
+        if product.user_id!=current_user.id:
+            return jsonify(msg="Unauthorized request!"), 403
+        
         db.session.delete(product)
         db.session.commit()
 
-        return jsonify({'result': True}), 200
+        return jsonify(msg="Successfully deleted product!"), 200
     except:
         db.session.rollback()
-        return "Error occurred", 500
+        return jsonify(msg="Error occurred!"), 500
     
 @products_bp.put('/edit/<int:product_id>')
-@login_required
 @jwt_required()
 def updateProduct(product_id):
-    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar()
+    product=db.session.execute(select(Product).where(Product.id==product_id)).scalar_one_or_none()
 
     try:
+        if product.user_id!=current_user.id:
+            return jsonify(msg="Unauthorized request!"), 403
+        
         data = request.get_json()
 
         if not (data['name'] == ''):
@@ -92,7 +90,7 @@ def updateProduct(product_id):
 
         db.session.commit()
 
-        return jsonify({'result': True}), 200
+        return jsonify(msg="Succesfully updated product!"), 201
     except:
         db.session.rollback()
-        return "Error occurred", 500
+        return jsonify(msg="Error occurred!"), 500
